INTG-1687 INTG-1662 Categorize obsolete data structure
