+++ Runtime/Scripts/Data/Packets/PacketDeserializer.cs
@@ -8,7 +8,6 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
-using UnityEngine;
 
 namespace Inworld.Packet
 {
