+++ Runtime/Scripts/Deprecated/DeprecatedData.cs
@@ -1,0 +1,71 @@
+﻿/*************************************************************************************************
+ * Copyright 2022-2024 Theai, Inc. dba Inworld AI
+ *
+ * Use of this source code is governed by the Inworld.ai Software Development Kit License Agreement
+ * that can be found in the LICENSE.md file or at https://www.inworld.ai/sdk-license
+ *************************************************************************************************/
+using System;
+using System.Collections.Generic;
+using UnityEngine;
+namespace Inworld.Packet
+{
+    [Serializable][Obsolete]
+    public class GestureEvent
+    {
+        public string type;
+        public string playback;
+    }
+    [Serializable][Obsolete]
+    public class GesturePacket : InworldPacket
+    {
+        public GestureEvent gesture;
+        
+        public GesturePacket()
+        {
+            type = PacketType.GESTURE;
+            gesture = new GestureEvent();
+        }
+        public GesturePacket(InworldPacket rhs, GestureEvent evt) : base(rhs)
+        {
+            type = PacketType.GESTURE;
+            gesture = evt;
+        }
+    }
+    [Obsolete]
+    public class LoadScenePacket : InworldPacket
+    {
+        public LoadSceneEvent mutation;
+
+        public LoadScenePacket(string sceneFullName)
+        {
+            timestamp = InworldDateTime.UtcNow;
+            type = PacketType.MUTATION;
+            packetId = new PacketId();
+            routing = new Routing("WORLD");
+            mutation = new LoadSceneEvent
+            {
+                loadScene = new LoadSceneRequest
+                {
+                    name = sceneFullName
+                }
+            };
+        }
+    }
+    [Obsolete]
+    public class LoadCharactersPacket : InworldPacket
+    {
+        public LoadCharactersEvent mutation;
+
+        public LoadCharactersPacket(List<string> characterFullName)
+        {
+            timestamp = InworldDateTime.UtcNow;
+            type = PacketType.MUTATION;
+            packetId = new PacketId();
+            routing = new Routing();
+            mutation = new LoadCharactersEvent
+            {
+                loadCharacters = new LoadCharactersRequest(characterFullName)
+            };
+        }
+    }
+}
