+++ Runtime/Scripts/Deprecated/InworldWebSocketClient.cs
@@ -5,8 +5,10 @@
  * that can be found in the LICENSE.md file or at https://www.inworld.ai/sdk-license
  *************************************************************************************************/
 
+using System;
 namespace Inworld
 {
+	[Obsolete]
     public class InworldWebSocketClient : InworldClient
     {
