+++ Runtime/Scripts/Deprecated/SessionResponsePacket.cs
